py_requirements.txt
numpy==1.23.3
pygame==2.1.2
PyOpenGL==3.1.6
pyrr==0.10.3
pyserial==3.5